Started openai gym Hanabi environemnt.
[OpenAI Gym][gym] is a library of common reinforcement learning
environments. [rllab][rllab] is a library of common reinforcement
learning algorithms. The algorithms in rllab can be used to learn the
environments in gym. This commit starts to implement Hanabi as a gym
environment following the guidelines [here][env].

Gym environments represent actions and observations as Spaces (e.g.
Discrete, Box, etc.). This commit implements the basic components of
Hanabi and provides functions to map them to and from samples of gym
spaces.

Next up, we have to implement the game state, implement the environment,
and implement a bot.

[env]: https://github.com/openai/gym/blob/master/gym/envs/README.md
[gym]: https://gym.openai.com/
[rllab]: https://github.com/openai/rllab