Tidied up HanabiAiEnv and HanabiSelfEnv.
We noticed that rllab was crashing when we returned non-empty info
dicts. We have no idea why; it could simply be a bug in rllab. Until we
figure out why, we return an empty info dict. We're currently not using
the info for much anyway, so it's not a big deal.